feat: remove region/route templates
